refactor: 모델 도메인 분리(#165)
