+++ products/models/__init__.py
@@ -1,0 +1,3 @@
+from .product import *
+from .category import *
+from .ingredient import *
